Update and rename Invalid-Identifier-Characters.CS to InvalidIdentifierCharacterMapAndLookupMemoryCache.CS
Haven't ran this sucka-yet, but the lips are wet; and everything is error free thus far. Should initialize runtime-continuity and default the lookup table. Great usage example for Locales and such volatile mapping systems. ASCII is -127 - 128, and shifting these values to incorporate seems to leak foreign characters in. Which may be of use, I seem to, regardless, only receive a limited charset; but am able to slide amongst into and out of the normal delineated charmap for the system; apparently there are TONS of characters I may use!